Implement cleanup/ backup-file-name option features
Cleanup will now be performed after backup. So only the specified amount of backup files per database will be kept, the oldest will be deleted.

It can be specified to use Database name or File name for Backup File Name and wether to use recycle bin or not for removed backup files.