Simplify logger and add option to log to file
The logger does not need to implement the KeePass
IStatusLogger interface, as it's only used by the
plugin to log some internal plugin actions.
Additionally added the feature to log the messages
to a log file. This has to be enabled "manually"
by setting the corresponsing config value in the
"KeePass.config.xml" file.

See #17