Move handler for closing database to "file closing post"
Instead of handling a closed database in the "File-Saving-Pre"
event, it is now handled in the "File-Saving-Post" event.
This ensures, that the normal database save (by KeePass) is
triggered and performed before the backup plugin kicks in
(because the "regular" save is more important than the backup,
it should be performed first).
The deregistration of the handler in the Terminate() function was
already on the correct event.