Copy LTB backup files instead of exporting each separatly
Instead of exporting (i.e., saving through the KeePass API), the
LTB backup files are copied from only one exported file.
This is either created as temporary file, or - if available - the backup
file of the normal backup is used for copying.

To ensure compatiblity with "external"/synced storages the original
file is not copied. Instead, at least one file is saved through the API.
(Either the normal backup, or a temporary export)

see #28